Fix and refactor get_parse_fixtures
