+++ test/fixtures/parser/ansi/create_database_a.yml
@@ -3,8 +3,5 @@ file:
     create_database_statement:
     - keyword: create
     - keyword: database
-    - keyword: if
-    - keyword: not
-    - keyword: exists
     - database_reference:
         identifier: my_database
