+++ test/fixtures/parser/postgres/postgres_create_extension.yml
@@ -16,4 +16,3 @@ file:
     - identifier: '2.0'
     - keyword: FROM
     - identifier: '1.0'
-
